+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/productredesign/ui/create/OpenFoodFactsError.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.productredesign.ui
+package com.maksimowiczm.foodyou.feature.productredesign.ui.create
 
 internal sealed interface OpenFoodFactsError {
     data object InvalidUrl : OpenFoodFactsError
